change api!: moved `max_edge_order()`, `is_possible_order()`, and `is_uniform()` to function.py
