+++ tests/classes/test_hypergraph.py
@@ -97,21 +97,6 @@ def test_dual(edgelist1, edgelist2, edgelist4):
     assert (D3.num_nodes, D3.num_edges) == (3, 5)
 
 
-def test_max_edge_order(edgelist1, edgelist4, edgelist5):
-    H0 = xgi.empty_hypergraph()
-    H1 = xgi.empty_hypergraph()
-    H1.add_nodes_from(range(5))
-    H2 = xgi.Hypergraph(edgelist1)
-    H3 = xgi.Hypergraph(edgelist4)
-    H4 = xgi.Hypergraph(edgelist5)
-
-    assert H0.max_edge_order() == None
-    assert H1.max_edge_order() == 0
-    assert H2.max_edge_order() == 2
-    assert H3.max_edge_order() == 3
-    assert H4.max_edge_order() == 3
-
-
 def test_add_nodes_from(attr1, attr2, attr3):
     H = xgi.Hypergraph()
     H.add_nodes_from(range(3), **attr1)
@@ -124,16 +109,6 @@ def test_add_nodes_from(attr1, attr2, attr3):
     assert H.nodes[0]["color"] == attr1["color"]
     assert H.nodes[1]["color"] == attr2["color"]
     assert H.nodes[2]["color"] == attr3["color"]
-
-
-def test_is_possible_order(edgelist1):
-    H1 = xgi.Hypergraph(edgelist1)
-
-    assert H1.is_possible_order(-1) == False
-    assert H1.is_possible_order(0) == False
-    assert H1.is_possible_order(1) == True
-    assert H1.is_possible_order(2) == True
-    assert H1.is_possible_order(3) == False
 
 
 def test_singleton_edges(edgelist1, edgelist2):
@@ -154,18 +129,6 @@ def test_remove_singleton_edges(edgelist1, edgelist2):
 
     assert H1.singleton_edges() == {}
     assert H2.singleton_edges() == {}
-
-
-def test_is_uniform(edgelist1, edgelist6, edgelist7):
-    H0 = xgi.Hypergraph(edgelist1)
-    H1 = xgi.Hypergraph(edgelist6)
-    H2 = xgi.Hypergraph(edgelist7)
-    H3 = xgi.empty_hypergraph()
-
-    assert H0.is_uniform() == False
-    assert H1.is_uniform() == 2
-    assert H2.is_uniform() == 2
-    assert H3.is_uniform() == False
 
 
 def test_isolates(edgelist1):
